// src/models/types.d.ts
export type ValueType = "NUMERIC" | "BOOLEAN" | "TEXT";
export type FeatureType = "INFORMATION" | "INTEGRATION" | "DOMAIN" | "AUTOMATION" | "MANAGEMENT" | "GUARANTEE" | "SUPPORT" | "PAYMENT";
export type UsageLimitType = "RENEWABLE" | "NON_RENEWABLE" | "TIME_DRIVEN" | "RESPONSE_DRIVEN";
export type RenderMode = "auto" | "enabled" | "disabled";